requirements.txt
python-telegram-bot
python-telegram-bot[job-queue]
requests
httpx
pytz
python-dotenv
dotwiz